Convert JavaScript integration to provider arrays
